CORE-17274 Remove Backchain resolution flow 5.0 compatibility mode
