--- components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/flows/backchain/TransactionBackChainResolutionVersion.kt
@@ -1,6 +1,0 @@
-package net.corda.ledger.utxo.flow.impl.flows.backchain
-
-enum class TransactionBackChainResolutionVersion {
-    V1, //  < 5.1
-    V2  // >= 5.1
-}
