+++ components/ledger/ledger-utxo-flow/src/test/kotlin/net/corda/ledger/utxo/flow/impl/flows/backchain/TransactionBackchainSenderFlowVersionedFlowFactoryTest.kt
@@ -2,6 +2,7 @@ package net.corda.ledger.utxo.flow.impl.flows.backchain
 
 import net.corda.ledger.utxo.flow.impl.flows.backchain.v1.TransactionBackchainSenderFlowV1
 import net.corda.libs.platform.PlatformVersion.CORDA_5_1
+import net.corda.v5.base.exceptions.CordaRuntimeException
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
@@ -13,43 +14,30 @@ class TransactionBackchainSenderFlowVersionedFlowFactoryTest {
     private val factory = TransactionBackchainSenderFlowVersionedFlowFactory(mock())
 
     @Test
-    fun `with platform version 1 creates TransactionBackchainSenderFlowV1`() {
-        val flow = factory.create(1, listOf(mock()))
-        assertThat(flow).isExactlyInstanceOf(TransactionBackchainSenderFlowV1::class.java)
-        assertSame(
-            TransactionBackChainResolutionVersion.V1,
-            (flow as TransactionBackchainSenderFlowV1).version
-        )
+    fun `with platform version 1 throws a CordaRuntimeException`() {
+        assertThatThrownBy { factory.create(1, listOf(mock())) }.isInstanceOf(CordaRuntimeException::class.java)
     }
 
     @Test
-    fun `with last potential 5_0 platform version creates TransactionBackchainSenderFlowV1`() {
-        val flow = factory.create(CORDA_5_1.value - 1, listOf(mock()))
-        assertThat(flow).isExactlyInstanceOf(TransactionBackchainSenderFlowV1::class.java)
-        assertSame(
-            TransactionBackChainResolutionVersion.V1,
-            (flow as TransactionBackchainSenderFlowV1).version
-        )
+    fun `with last potential 5_0 platform version throws a CordaRuntimeException`() {
+        assertThatThrownBy {
+            factory.create(
+                CORDA_5_1.value - 1,
+                listOf(mock())
+            )
+        }.isInstanceOf(CordaRuntimeException::class.java)
     }
 
     @Test
-    fun `with first 5_1 platform version creates TransactionBackchainSenderFlowV2`() {
+    fun `with first 5_1 platform version creates TransactionBackchainSenderFlowV1`() {
         val flow = factory.create(CORDA_5_1.value, listOf(mock()))
         assertThat(flow).isExactlyInstanceOf(TransactionBackchainSenderFlowV1::class.java)
-        assertSame(
-            TransactionBackChainResolutionVersion.V2,
-            (flow as TransactionBackchainSenderFlowV1).version
-        )
     }
 
     @Test
-    fun `with platform version 50199 creates TransactionBackchainSenderFlowV2`() {
+    fun `with platform version 50199 creates TransactionBackchainSenderFlowV1`() {
         val flow = factory.create(50199, listOf(mock()))
         assertThat(flow).isExactlyInstanceOf(TransactionBackchainSenderFlowV1::class.java)
-        assertSame(
-            TransactionBackChainResolutionVersion.V2,
-            (flow as TransactionBackchainSenderFlowV1).version
-        )
     }
 
     @Test
