+++ components/ledger/ledger-utxo-flow/src/test/kotlin/net/corda/ledger/utxo/flow/impl/flows/backchain/v1/TransactionBackchainReceiverFlowV1Test.kt
@@ -3,24 +3,28 @@ package net.corda.ledger.utxo.flow.impl.flows.backchain.v1
 import net.corda.crypto.core.SecureHashImpl
 import net.corda.flow.application.services.FlowConfigService
 import net.corda.ledger.common.data.transaction.CordaPackageSummaryImpl
+import net.corda.ledger.common.data.transaction.TransactionMetadataInternal
 import net.corda.ledger.common.data.transaction.TransactionStatus.UNVERIFIED
 import net.corda.ledger.utxo.flow.impl.UtxoLedgerMetricRecorder
 import net.corda.ledger.utxo.flow.impl.flows.backchain.TopologicalSort
-import net.corda.ledger.utxo.flow.impl.flows.backchain.TransactionBackChainResolutionVersion
 import net.corda.ledger.utxo.flow.impl.groupparameters.verifier.SignedGroupParametersVerifier
 import net.corda.ledger.utxo.flow.impl.persistence.TransactionExistenceStatus
 import net.corda.ledger.utxo.flow.impl.persistence.UtxoLedgerGroupParametersPersistenceService
 import net.corda.ledger.utxo.flow.impl.persistence.UtxoLedgerPersistenceService
 import net.corda.libs.configuration.SmartConfig
+import net.corda.membership.lib.SignedGroupParameters
 import net.corda.schema.configuration.ConfigKeys
 import net.corda.v5.application.messaging.FlowSession
+import net.corda.v5.base.exceptions.CordaRuntimeException
 import net.corda.v5.crypto.SecureHash
+import net.corda.v5.crypto.exceptions.CryptoSignatureException
 import net.corda.v5.ledger.utxo.StateRef
 import net.corda.v5.ledger.utxo.transaction.UtxoSignedTransaction
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.mockito.Mockito.times
 import org.mockito.kotlin.any
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.eq
@@ -53,6 +57,13 @@ class TransactionBackchainReceiverFlowV1Test {
         const val BACKCHAIN_BATCH_DEFAULT_SIZE = 1
     }
 
+    private val groupParameters = mock<SignedGroupParameters>()
+    private val groupParametersHash1 = SecureHashImpl("SHA", byteArrayOf(101, 101, 101, 101))
+
+    private val tx1Metadata = mock<TransactionMetadataInternal>()
+    private val tx2Metadata = mock<TransactionMetadataInternal>()
+    private val tx3Metadata = mock<TransactionMetadataInternal>()
+
     private val utxoLedgerPersistenceService = mock<UtxoLedgerPersistenceService>()
     private val utxoLedgerMetricRecorder = mock<UtxoLedgerMetricRecorder>()
     private val utxoLedgerGroupParametersPersistenceService = mock<UtxoLedgerGroupParametersPersistenceService>()
@@ -83,35 +94,42 @@ class TransactionBackchainReceiverFlowV1Test {
             listOf(retrievedTransaction3)
         )
 
+        whenever(session.sendAndReceive(eq(SignedGroupParameters::class.java), any())).thenReturn(
+            groupParameters,
+        )
+        whenever(groupParameters.hash).thenReturn(groupParametersHash1)
+
         whenever(utxoLedgerPersistenceService.persistIfDoesNotExist(any(), eq(UNVERIFIED)))
             .thenReturn(TransactionExistenceStatus.DOES_NOT_EXIST to listOf(PACKAGE_SUMMARY))
 
         whenever(retrievedTransaction1.id).thenReturn(TX_ID_1)
         whenever(retrievedTransaction1.inputStateRefs).thenReturn(listOf(TX_3_INPUT_DEPENDENCY_STATE_REF_1))
         whenever(retrievedTransaction1.referenceStateRefs).thenReturn(listOf(TX_3_INPUT_REFERENCE_DEPENDENCY_STATE_REF_1))
+        whenever(retrievedTransaction1.metadata).thenReturn(tx1Metadata)
+        whenever(tx1Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         whenever(retrievedTransaction2.id).thenReturn(TX_ID_2)
         whenever(retrievedTransaction2.inputStateRefs).thenReturn(listOf(TX_3_INPUT_DEPENDENCY_STATE_REF_2))
         whenever(retrievedTransaction2.referenceStateRefs).thenReturn(listOf(TX_3_INPUT_REFERENCE_DEPENDENCY_STATE_REF_2))
+        whenever(retrievedTransaction2.metadata).thenReturn(tx1Metadata)
+        whenever(tx2Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         whenever(retrievedTransaction3.id).thenReturn(TX_ID_3)
         whenever(retrievedTransaction3.inputStateRefs).thenReturn(emptyList())
         whenever(retrievedTransaction3.referenceStateRefs).thenReturn(emptyList())
+        whenever(retrievedTransaction3.metadata).thenReturn(tx1Metadata)
+        whenever(tx3Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         assertThat(callTransactionBackchainReceiverFlow(setOf(TX_ID_1, TX_ID_2)).complete()).isEqualTo(listOf(TX_ID_3, TX_ID_2, TX_ID_1))
 
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_1)))
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_2)))
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_3)))
+        verify(session, times(3)).sendAndReceive(SignedGroupParameters::class.java, TransactionBackchainRequestV1.GetSignedGroupParameters(groupParametersHash1))
         verify(session).send(TransactionBackchainRequestV1.Stop)
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction1, UNVERIFIED)
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction2, UNVERIFIED)
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction3, UNVERIFIED)
-
-        verifyNoInteractions(
-            utxoLedgerGroupParametersPersistenceService,
-            signedGroupParametersVerifier
-        )
     }
 
     @Test
@@ -136,33 +154,42 @@ class TransactionBackchainReceiverFlowV1Test {
             listOf(retrievedTransaction3)
         )
 
+        whenever(session.sendAndReceive(eq(SignedGroupParameters::class.java), any())).thenReturn(
+            groupParameters,
+        )
+        whenever(groupParameters.hash).thenReturn(groupParametersHash1)
+
         whenever(utxoLedgerPersistenceService.persistIfDoesNotExist(any(), eq(UNVERIFIED)))
             .thenReturn(TransactionExistenceStatus.DOES_NOT_EXIST to listOf(PACKAGE_SUMMARY))
 
         whenever(retrievedTransaction1.id).thenReturn(TX_ID_1)
         whenever(retrievedTransaction1.inputStateRefs).thenReturn(listOf(TX_3_INPUT_DEPENDENCY_STATE_REF_1))
         whenever(retrievedTransaction1.referenceStateRefs).thenReturn(listOf(TX_3_INPUT_REFERENCE_DEPENDENCY_STATE_REF_1))
+        whenever(retrievedTransaction1.metadata).thenReturn(tx1Metadata)
+        whenever(tx1Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         whenever(retrievedTransaction2.id).thenReturn(TX_ID_2)
         whenever(retrievedTransaction2.inputStateRefs).thenReturn(listOf(TX_3_INPUT_DEPENDENCY_STATE_REF_2))
         whenever(retrievedTransaction2.referenceStateRefs).thenReturn(listOf(TX_3_INPUT_REFERENCE_DEPENDENCY_STATE_REF_2))
+        whenever(retrievedTransaction2.metadata).thenReturn(tx1Metadata)
+        whenever(tx2Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         whenever(retrievedTransaction3.id).thenReturn(TX_ID_3)
         whenever(retrievedTransaction3.inputStateRefs).thenReturn(emptyList())
         whenever(retrievedTransaction3.referenceStateRefs).thenReturn(emptyList())
+        whenever(retrievedTransaction3.metadata).thenReturn(tx1Metadata)
+        whenever(tx3Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         assertThat(callTransactionBackchainReceiverFlow(setOf(TX_ID_1, TX_ID_2)).complete()).isEqualTo(listOf(TX_ID_3, TX_ID_2, TX_ID_1))
 
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_1)))
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_2)))
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_3)))
+        verify(session, times(3)).sendAndReceive(SignedGroupParameters::class.java, TransactionBackchainRequestV1.GetSignedGroupParameters(groupParametersHash1))
+        verify(session).send(TransactionBackchainRequestV1.Stop)
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction1, UNVERIFIED)
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction2, UNVERIFIED)
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction3, UNVERIFIED)
-        verifyNoInteractions(
-            utxoLedgerGroupParametersPersistenceService,
-            signedGroupParametersVerifier
-        )
     }
 
     @Test
@@ -174,15 +201,24 @@ class TransactionBackchainReceiverFlowV1Test {
             listOf(retrievedTransaction2)
         )
 
+        whenever(session.sendAndReceive(eq(SignedGroupParameters::class.java), any())).thenReturn(
+            groupParameters,
+        )
+        whenever(groupParameters.hash).thenReturn(groupParametersHash1)
+
         whenever(utxoLedgerPersistenceService.persistIfDoesNotExist(any(), eq(UNVERIFIED)))
             .thenReturn(TransactionExistenceStatus.DOES_NOT_EXIST to listOf(PACKAGE_SUMMARY))
 
         whenever(utxoLedgerPersistenceService.persistIfDoesNotExist(retrievedTransaction1, UNVERIFIED))
             .thenReturn(TransactionExistenceStatus.VERIFIED to listOf(PACKAGE_SUMMARY))
+        whenever(retrievedTransaction1.metadata).thenReturn(tx1Metadata)
+        whenever(tx1Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         whenever(retrievedTransaction1.id).thenReturn(TX_ID_1)
         whenever(retrievedTransaction1.inputStateRefs).thenReturn(listOf(TX_3_INPUT_DEPENDENCY_STATE_REF_1))
         whenever(retrievedTransaction1.referenceStateRefs).thenReturn(listOf(TX_3_INPUT_REFERENCE_DEPENDENCY_STATE_REF_1))
+        whenever(retrievedTransaction2.metadata).thenReturn(tx1Metadata)
+        whenever(tx2Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         whenever(retrievedTransaction2.id).thenReturn(TX_ID_2)
         whenever(retrievedTransaction2.inputStateRefs).thenReturn(emptyList())
@@ -193,33 +229,35 @@ class TransactionBackchainReceiverFlowV1Test {
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_1)))
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_2)))
         verify(session, never()).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_3)))
+        verify(session, times(2)).sendAndReceive(SignedGroupParameters::class.java, TransactionBackchainRequestV1.GetSignedGroupParameters(groupParametersHash1))
+        verify(session).send(TransactionBackchainRequestV1.Stop)
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction1, UNVERIFIED)
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction2, UNVERIFIED)
         verify(utxoLedgerPersistenceService, never()).persistIfDoesNotExist(retrievedTransaction3, UNVERIFIED)
-        verifyNoInteractions(
-            utxoLedgerGroupParametersPersistenceService,
-            signedGroupParametersVerifier
-        )
     }
 
     @Test
     fun `receiving only transactions that are stored locally as VERIFIED does not have their dependencies added to the transactions to retrieve and stops resolution`() {
-        whenever(utxoLedgerPersistenceService.findSignedTransaction(TX_ID_1)).thenReturn(retrievedTransaction1)
-
         whenever(session.sendAndReceive(eq(List::class.java), any())).thenReturn(
             listOf(retrievedTransaction1),
             listOf(retrievedTransaction2)
         )
 
-        whenever(utxoLedgerPersistenceService.persistIfDoesNotExist(any(), eq(UNVERIFIED)))
-            .thenReturn(TransactionExistenceStatus.DOES_NOT_EXIST to listOf(PACKAGE_SUMMARY))
+        whenever(session.sendAndReceive(eq(SignedGroupParameters::class.java), any())).thenReturn(
+            groupParameters,
+        )
+        whenever(groupParameters.hash).thenReturn(groupParametersHash1)
 
         whenever(utxoLedgerPersistenceService.persistIfDoesNotExist(any(), eq(UNVERIFIED)))
             .thenReturn(TransactionExistenceStatus.VERIFIED to listOf(PACKAGE_SUMMARY))
+        whenever(retrievedTransaction1.metadata).thenReturn(tx1Metadata)
+        whenever(tx1Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         whenever(retrievedTransaction1.id).thenReturn(TX_ID_1)
         whenever(retrievedTransaction1.inputStateRefs).thenReturn(listOf(TX_3_INPUT_DEPENDENCY_STATE_REF_1))
         whenever(retrievedTransaction1.referenceStateRefs).thenReturn(listOf(TX_3_INPUT_REFERENCE_DEPENDENCY_STATE_REF_1))
+        whenever(retrievedTransaction2.metadata).thenReturn(tx1Metadata)
+        whenever(tx2Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         whenever(retrievedTransaction2.id).thenReturn(TX_ID_2)
         whenever(retrievedTransaction2.inputStateRefs).thenReturn(emptyList())
@@ -230,14 +268,11 @@ class TransactionBackchainReceiverFlowV1Test {
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_1)))
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_2)))
         verify(session, never()).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_3)))
+        verify(session, times(2)).sendAndReceive(SignedGroupParameters::class.java, TransactionBackchainRequestV1.GetSignedGroupParameters(groupParametersHash1))
         verify(session).send(TransactionBackchainRequestV1.Stop)
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction1, UNVERIFIED)
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction2, UNVERIFIED)
         verify(utxoLedgerPersistenceService, never()).persistIfDoesNotExist(retrievedTransaction3, UNVERIFIED)
-        verifyNoInteractions(
-            utxoLedgerGroupParametersPersistenceService,
-            signedGroupParametersVerifier
-        )
     }
 
     @Test
@@ -249,12 +284,19 @@ class TransactionBackchainReceiverFlowV1Test {
             listOf(retrievedTransaction2)
         )
 
+        whenever(session.sendAndReceive(eq(SignedGroupParameters::class.java), any())).thenReturn(
+            groupParameters,
+        )
+        whenever(groupParameters.hash).thenReturn(groupParametersHash1)
+
         whenever(utxoLedgerPersistenceService.persistIfDoesNotExist(retrievedTransaction1, UNVERIFIED))
             .thenReturn(TransactionExistenceStatus.DOES_NOT_EXIST to listOf(PACKAGE_SUMMARY))
 
         whenever(retrievedTransaction1.id).thenReturn(TX_ID_1)
         whenever(retrievedTransaction1.inputStateRefs).thenReturn(listOf(TX_3_INPUT_DEPENDENCY_STATE_REF_1))
         whenever(retrievedTransaction1.referenceStateRefs).thenReturn(listOf(TX_3_INPUT_REFERENCE_DEPENDENCY_STATE_REF_1))
+        whenever(retrievedTransaction1.metadata).thenReturn(tx1Metadata)
+        whenever(tx1Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         whenever(retrievedTransaction2.id).thenReturn(TX_ID_2)
 
@@ -264,10 +306,90 @@ class TransactionBackchainReceiverFlowV1Test {
         verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_1)))
         verify(utxoLedgerPersistenceService).persistIfDoesNotExist(retrievedTransaction1, UNVERIFIED)
         verify(utxoLedgerPersistenceService, never()).persistIfDoesNotExist(retrievedTransaction2, UNVERIFIED)
-        verifyNoInteractions(
-            utxoLedgerGroupParametersPersistenceService,
-            signedGroupParametersVerifier
+    }
+
+    @Test
+    fun `receiving signed group parameters that was not requested  throws an exception`() {
+        whenever(utxoLedgerPersistenceService.findSignedTransaction(TX_ID_1)).thenReturn(retrievedTransaction1)
+
+        whenever(session.sendAndReceive(eq(List::class.java), any())).thenReturn(
+            listOf(retrievedTransaction1),
         )
+
+        whenever(session.sendAndReceive(eq(SignedGroupParameters::class.java), any())).thenReturn(
+            groupParameters,
+        )
+        whenever(groupParameters.hash).thenReturn(SecureHashImpl("SHA", byteArrayOf(103, 104, 105, 106)))
+
+        whenever(utxoLedgerPersistenceService.persistIfDoesNotExist(retrievedTransaction1, UNVERIFIED))
+            .thenReturn(TransactionExistenceStatus.DOES_NOT_EXIST to listOf(PACKAGE_SUMMARY))
+
+        whenever(retrievedTransaction1.id).thenReturn(TX_ID_1)
+        whenever(retrievedTransaction1.inputStateRefs).thenReturn(listOf(TX_3_INPUT_DEPENDENCY_STATE_REF_1))
+        whenever(retrievedTransaction1.referenceStateRefs).thenReturn(listOf(TX_3_INPUT_REFERENCE_DEPENDENCY_STATE_REF_1))
+        whenever(retrievedTransaction1.metadata).thenReturn(tx1Metadata)
+        whenever(tx1Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
+
+        assertThatThrownBy { callTransactionBackchainReceiverFlow(setOf(TX_ID_1)) }
+            .isExactlyInstanceOf(CordaRuntimeException::class.java)
+            .hasMessageContaining("but received:")
+
+        verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_1)))
+        verify(utxoLedgerPersistenceService, never()).persistIfDoesNotExist(eq(retrievedTransaction2), any())
+    }
+
+    @Test
+    fun `receiving signed group parameters with invalid signature throws an exception`() {
+        whenever(utxoLedgerPersistenceService.findSignedTransaction(TX_ID_1)).thenReturn(retrievedTransaction1)
+
+        whenever(session.sendAndReceive(eq(List::class.java), any())).thenReturn(
+            listOf(retrievedTransaction1),
+        )
+
+        whenever(session.sendAndReceive(eq(SignedGroupParameters::class.java), any())).thenReturn(
+            groupParameters,
+        )
+        whenever(groupParameters.hash).thenReturn(groupParametersHash1)
+        whenever(signedGroupParametersVerifier.verifySignature(any())).thenThrow(
+            CryptoSignatureException("Invalid signature")
+        )
+
+        whenever(utxoLedgerPersistenceService.persistIfDoesNotExist(retrievedTransaction1, UNVERIFIED))
+            .thenReturn(TransactionExistenceStatus.DOES_NOT_EXIST to listOf(PACKAGE_SUMMARY))
+
+        whenever(retrievedTransaction1.id).thenReturn(TX_ID_1)
+        whenever(retrievedTransaction1.inputStateRefs).thenReturn(listOf(TX_3_INPUT_DEPENDENCY_STATE_REF_1))
+        whenever(retrievedTransaction1.referenceStateRefs).thenReturn(listOf(TX_3_INPUT_REFERENCE_DEPENDENCY_STATE_REF_1))
+        whenever(retrievedTransaction1.metadata).thenReturn(tx1Metadata)
+        whenever(tx1Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
+
+        assertThatThrownBy { callTransactionBackchainReceiverFlow(setOf(TX_ID_1)) }
+            .isExactlyInstanceOf(CryptoSignatureException::class.java)
+            .hasMessageContaining("Invalid signature")
+
+        verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_1)))
+        verify(utxoLedgerPersistenceService, never()).persistIfDoesNotExist(eq(retrievedTransaction2), any())
+    }
+
+    @Test
+    fun `receiving a transaction without signed group parameters hash in its metadata throws an exception`() {
+        whenever(utxoLedgerPersistenceService.findSignedTransaction(TX_ID_1)).thenReturn(retrievedTransaction1)
+
+        whenever(session.sendAndReceive(eq(List::class.java), any())).thenReturn(
+            listOf(retrievedTransaction1),
+        )
+        whenever(groupParameters.hash).thenReturn(groupParametersHash1)
+
+        whenever(retrievedTransaction1.id).thenReturn(TX_ID_1)
+        whenever(retrievedTransaction1.metadata).thenReturn(tx1Metadata)
+        whenever(tx1Metadata.getMembershipGroupParametersHash()).thenReturn(null)
+
+
+        assertThatThrownBy { callTransactionBackchainReceiverFlow(setOf(TX_ID_1)) }
+            .isExactlyInstanceOf(IllegalArgumentException::class.java)
+
+        verify(session).sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(TX_ID_1)))
+        verify(utxoLedgerPersistenceService, never()).persistIfDoesNotExist(eq(retrievedTransaction1), any())
     }
 
     @Test
@@ -292,6 +414,26 @@ class TransactionBackchainReceiverFlowV1Test {
 
        TX5 is not referenced in the test because the dependencies of the transaction are passed into the flow as IDs.
        */
+        val groupParameters4 = mock<SignedGroupParameters>()
+        val groupParameters3 = mock<SignedGroupParameters>()
+        val groupParameters2 = mock<SignedGroupParameters>()
+        val groupParameters1 = mock<SignedGroupParameters>()
+
+        val groupParametersHash4 = SecureHashImpl("SHA", byteArrayOf(104, 104, 104, 104))
+        val groupParametersHash3 = SecureHashImpl("SHA", byteArrayOf(103, 103, 103, 103))
+        val groupParametersHash2 = SecureHashImpl("SHA", byteArrayOf(102, 102, 102, 102))
+        val groupParametersHash1 = SecureHashImpl("SHA", byteArrayOf(101, 101, 101, 101))
+
+        whenever(groupParameters4.hash).thenReturn(groupParametersHash4)
+        whenever(groupParameters3.hash).thenReturn(groupParametersHash3)
+        whenever(groupParameters2.hash).thenReturn(groupParametersHash2)
+        whenever(groupParameters1.hash).thenReturn(groupParametersHash1)
+
+
+        val tx4Metadata = mock<TransactionMetadataInternal>()
+        val tx3Metadata = mock<TransactionMetadataInternal>()
+        val tx2Metadata = mock<TransactionMetadataInternal>()
+        val tx1Metadata = mock<TransactionMetadataInternal>()
 
         val transactionId4 = SecureHashImpl("SHA", byteArrayOf(4, 4, 4, 4))
         val transactionId3 = SecureHashImpl("SHA", byteArrayOf(3, 3, 3, 3))
@@ -309,15 +451,24 @@ class TransactionBackchainReceiverFlowV1Test {
 
         whenever(transaction4.id).thenReturn(transactionId4)
         whenever(transaction4.inputStateRefs).thenReturn(listOf(transaction2StateRef))
+        whenever(transaction4.metadata).thenReturn(tx4Metadata)
 
         whenever(transaction3.id).thenReturn(transactionId3)
         whenever(transaction3.inputStateRefs).thenReturn(listOf(transaction1StateRef1))
+        whenever(transaction3.metadata).thenReturn(tx3Metadata)
 
         whenever(transaction2.id).thenReturn(transactionId2)
         whenever(transaction2.inputStateRefs).thenReturn(listOf(transaction1StateRef0))
+        whenever(transaction2.metadata).thenReturn(tx2Metadata)
 
         whenever(transaction1.id).thenReturn(transactionId1)
         whenever(transaction1.inputStateRefs).thenReturn(emptyList())
+        whenever(transaction1.metadata).thenReturn(tx1Metadata)
+
+        whenever(tx4Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash4.toString())
+        whenever(tx3Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash3.toString())
+        whenever(tx2Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash2.toString())
+        whenever(tx1Metadata.getMembershipGroupParametersHash()).thenReturn(groupParametersHash1.toString())
 
         whenever(utxoLedgerPersistenceService.findSignedTransaction(any(), any())).thenReturn(null)
 
@@ -326,6 +477,13 @@ class TransactionBackchainReceiverFlowV1Test {
             listOf(transaction4),
             listOf(transaction1),
             listOf(transaction2)
+        )
+
+        whenever(session.sendAndReceive(eq(SignedGroupParameters::class.java), any())).thenReturn(
+            groupParameters3,
+            groupParameters4,
+            groupParameters1,
+            groupParameters2,
         )
 
         whenever(utxoLedgerPersistenceService.persistIfDoesNotExist(any(), eq(UNVERIFIED)))
@@ -342,9 +500,13 @@ class TransactionBackchainReceiverFlowV1Test {
 
         session.inOrder {
             verify().sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(transactionId3)))
+            verify().sendAndReceive(SignedGroupParameters::class.java, TransactionBackchainRequestV1.GetSignedGroupParameters(groupParametersHash3))
             verify().sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(transactionId4)))
+            verify().sendAndReceive(SignedGroupParameters::class.java, TransactionBackchainRequestV1.GetSignedGroupParameters(groupParametersHash4))
             verify().sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(transactionId1)))
+            verify().sendAndReceive(SignedGroupParameters::class.java, TransactionBackchainRequestV1.GetSignedGroupParameters(groupParametersHash1))
             verify().sendAndReceive(List::class.java, TransactionBackchainRequestV1.Get(setOf(transactionId2)))
+            verify().sendAndReceive(SignedGroupParameters::class.java, TransactionBackchainRequestV1.GetSignedGroupParameters(groupParametersHash2))
             Unit
 
         }
@@ -356,17 +518,12 @@ class TransactionBackchainReceiverFlowV1Test {
             verify().persistIfDoesNotExist(transaction2, UNVERIFIED)
             Unit
         }
-        verifyNoInteractions(
-            utxoLedgerGroupParametersPersistenceService,
-            signedGroupParametersVerifier
-        )
     }
 
     private fun callTransactionBackchainReceiverFlow(originalTransactionsToRetrieve: Set<SecureHash>): TopologicalSort {
         return TransactionBackchainReceiverFlowV1(
             setOf(SecureHashImpl("SHA", byteArrayOf(1, 1, 1, 1))),
-            originalTransactionsToRetrieve, session,
-            TransactionBackChainResolutionVersion.V1
+            originalTransactionsToRetrieve, session
         ).apply {
             utxoLedgerPersistenceService = this@TransactionBackchainReceiverFlowV1Test.utxoLedgerPersistenceService
             utxoLedgerMetricRecorder = this@TransactionBackchainReceiverFlowV1Test.utxoLedgerMetricRecorder
